+++ pkg/providers/appsec/data_akamai_appsec_security_policy_protections.go
@@ -66,6 +66,11 @@ func dataSourcePolicyProtections() *schema.Resource {
 				Computed:    true,
 				Description: "Whether to enable slow post controls",
 			},
+			"apply_url_protection_controls": {
+				Type:        schema.TypeBool,
+				Computed:    true,
+				Description: "Whether to enable url protection controls",
+			},
 			"json": {
 				Type:        schema.TypeString,
 				Computed:    true,
@@ -158,6 +163,10 @@ func dataSourcePolicyProtectionsRead(ctx context.Context, d *schema.ResourceData
 	}
 
 	if err := d.Set("apply_slow_post_controls", policyprotections.ApplySlowPostControls); err != nil {
+		return diag.Errorf("%s: %s", tf.ErrValueSet, err.Error())
+	}
+
+	if err := d.Set("apply_url_protection_controls", policyprotections.ApplyURLProtectionControls); err != nil {
 		return diag.Errorf("%s: %s", tf.ErrValueSet, err.Error())
 	}